server/requirements.txt
openenv-core[core]>=0.2.3
orjson>=3.11.7
fastapi>=0.135.2
uvicorn>=0.42.0
